Mark the PackageDescription manifest API as being available only with tools version 999.0 since this hasn't been approved yet.
